backend/requirements.txt
numpy==1.24.2
# bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
Django==4.2.7
django-background-tasks==1.2.5
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-admin-rangefilter==0.9.0
django-render-block==0.9.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
gunicorn==20.1.0
lxml
pandas
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
urllib3==1.26.12
XlsxWriter==3.0.3
django-import-export==3.0.1
requests_oauthlib==1.3.1
python-gettext==4.0
django-environ==0.9.0
psycopg2-binary==2.9.9
pillow